+++ codex/serializers/browser.py
@@ -159,6 +159,7 @@ class BrowserSettingsSerializer(Serializer):
     order_reverse = BooleanField(required=False)
     q = CharField(allow_blank=True, required=False)
     show = BrowserSettingsShowGroupFlagsSerializer(required=False)
+    twenty_four_hour_time = BooleanField()
     top_group = ChoiceField(choices=tuple(CHOICES["topGroup"].keys()), required=False)